docs(recipe): remove DataSummaryCard recipe (#1532)
- this recipe, as-is, only shows how one might compose a stardard card, so it is redundant
- the layout in question is mostly specific to a single project, so not general purpose
- move a copy to use with the StudentRefinementPage
