+++ .storybook/pages/StudentRefinement/StudentRefinement.tsx
@@ -10,6 +10,7 @@ import {
   Checkbox,
   FiltersCheckboxField,
   FiltersDrawer,
+  Heading,
   Icon,
   Layout,
   PageHeader,
@@ -22,14 +23,103 @@ import {
   Tag,
   Text,
 } from '../../../src';
+import type { HeadingElement } from '../../../src/components/Heading';
 
 import breakpoint from '../../../src/design-tokens/tier-1-definitions/breakpoints';
 
 import { EdsThemeColorIconUtilityWarning } from '../../../src/tokens-dist/ts/colors';
 
-import { DataSummaryCard } from '../../recipes/DataSummaryCard/DataSummaryCard';
 import { PageShell } from '../../recipes/PageShell/PageShell';
 import styles from './StudentRefinement.module.css';
+
+interface DataSummaryProps {
+  /**
+   * CSS class names that can be appended to the component.
+   */
+  className?: string;
+  /**
+   * Main data text of the card.
+   */
+  dataAmount: string;
+  /**
+   * Text that provides a unit of measurement for the data.
+   */
+  dataUnit?: string;
+  /**
+   * Text to provide more context for the data.
+   */
+  description?: string;
+  /**
+   * Specifies the heading element to render the card heading as.
+   */
+  headingElement?: HeadingElement;
+  /**
+   * Title text of the data represented.
+   */
+  title: string;
+  /**
+   * Off track variant to indicate status.
+   */
+  variant?: 'off-track';
+}
+
+const DataSummaryCard = ({
+  className,
+  dataAmount,
+  dataUnit,
+  description,
+  headingElement,
+  title,
+  variant,
+  ...other
+}: DataSummaryProps) => {
+  const componentClassName = clsx(styles['data-summary-card'], className);
+  return (
+    <Card className={componentClassName} {...other}>
+      <Card.Header className={styles['data-summary-card__header']}>
+        <Heading
+          as={headingElement}
+          className={styles['data-summary-card__title']}
+          size="h3"
+        >
+          {title}
+        </Heading>
+      </Card.Header>
+      <Card.Body className={styles['data-summary-card__body']}>
+        <Text
+          className={styles['data-summary-card__data']}
+          variant="neutral-medium"
+        >
+          {dataAmount}
+          {dataUnit && (
+            <Text
+              as="span"
+              className={styles['data-summary-card__data-unit']}
+              variant="neutral-subtle"
+            >
+              {dataUnit}
+            </Text>
+          )}
+        </Text>
+        {description && (
+          <Text
+            className={styles['data-summary-card__description']}
+            variant="neutral-subtle"
+          >
+            {description}
+          </Text>
+        )}
+      </Card.Body>
+      {variant === 'off-track' && (
+        <div
+          aria-label="off track"
+          className={styles['data-summary-card__indicator--off-track']}
+          role="img"
+        />
+      )}
+    </Card>
+  );
+};
 
 export const StudentRefinement = () => {
   const [isTable, setIsTable] = useState(false);
