--- .storybook/recipes/DataSummaryCard/index.ts
@@ -1,1 +1,0 @@
-export { DataSummaryCard as default } from './DataSummaryCard';
